backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.17.2",
    "asyncpg>=0.31.0",
    "bcrypt>=5.0.0",
    "email-validator>=2.3.0",
    "fastapi>=0.124.2",
    "httpx>=0.28.0",
    "passlib[argon2]>=1.7.4",
    "pydantic-settings>=2.12.0",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.20",
    "redis>=7.1.0",
    "slowapi>=0.1.9",
    "sqlalchemy>=2.0.45",
    "uvicorn>=0.38.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.14.10",
]

[tool.ruff]
# Set the maximum line length to 100
line-length = 100
target-version = "py313"

# Exclude common directories
exclude = [
    ".git",
    ".venv",
    "__pycache__",
    "alembic/versions",
    ".mypy_cache",
    ".pytest_cache",
]

# Enable preview features for latest formatting improvements
preview = true

# Output format with upstream linter URLs
output-format = "concise"

[tool.ruff.lint]
# Enable these rule sets
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "N",      # pep8-naming
    "UP",     # pyupgrade
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "SIM",    # flake8-simplify
    "RET",    # flake8-return
    "ARG",    # flake8-unused-arguments
    "PTH",    # flake8-use-pathlib
    "ERA",    # eradicate (commented-out code)
    "PL",     # pylint
    "PERF",   # perflint
    "FURB",   # refurb (modernize Python code)
    "ISC",    # flake8-implicit-str-concat
]

# Ignore specific rules
ignore = [
    "E501",    # Line too long (handled by formatter)
    "PLR0913", # Too many arguments to function call
    "PLR2004", # Magic value used in comparison
    "ARG002",  # Unused method argument (often needed for interface compatibility)
    "PLR0912", # Too many branches (sometimes necessary)
    "B008",    # Function call in default argument (FastAPI Depends pattern)
    "UP042",   # Replace str+Enum with StrEnum (not available in all Python versions)
]

# Allow fix for all enabled rules (when `--fix` is provided)
fixable = ["ALL"]
unfixable = []

# Allow unused variables when underscore-prefixed
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.lint.per-file-ignores]
# Ignore import violations in __init__.py files
"__init__.py" = ["F401", "F403"]
# Ignore specific rules in migration files
"alembic/versions/*.py" = ["E501", "N999", "ARG001", "F401"]
# Ignore unused arguments in test files
"**/test_*.py" = ["ARG001", "ARG002"]
# Ignore top-level imports in main.py (FastAPI pattern) and unused args
"app/main.py" = ["PLC0415", "ARG001", "E402", "ERA001"]
# Ignore unused args in all API endpoint files (FastAPI pattern)
"app/api/v1/endpoints/*.py" = ["ARG001", "PLC0415", "B904", "E402"]
# Ignore complexity in booking/search endpoints (complex workflows)
"app/api/v1/endpoints/bookings.py" = ["PLR0917"]
"app/api/v1/endpoints/search.py" = ["PLR0917", "PLR0914"]
# Ignore StrEnum in schema files (compatibility)
"app/schemas/*.py" = ["UP042"]
# Ignore SCIM spec naming convention (mixedCase required by spec)
"app/schemas/scim.py" = ["N815"]
# Ignore staticmethod suggestions in service clients
"app/services/suppliers/*.py" = ["PLR6301", "B904"]
# Ignore F821 in all models (SQLAlchemy forward references)
"app/models/*.py" = ["F821", "PLC0415"]
# Ignore complexity warnings in services
"app/services/*.py" = ["PLR0917", "PLR0914", "PLW0603", "ARG003", "ARG004"]
# Ignore config comments
"app/core/config.py" = ["ERA001"]
# Ignore lazy import in transfer service
"app/services/transfer_service.py" = ["PLC0415"]
# Ignore exception handling in deps
"app/api/deps.py" = ["B904"]
# Ignore import order in booking schema (circular dependency)
"app/schemas/booking.py" = ["E402"]

[tool.ruff.format]
# Use double quotes for strings
quote-style = "double"

# Indent with spaces
indent-style = "space"

# Respect magic trailing comma
skip-magic-trailing-comma = false

# Automatically detect the appropriate line ending
line-ending = "auto"

# Preview: Improved lambda and method chain formatting
preview = true
